(PC-9610) Put the entire firebase analytics API behind our own interface
