+++ src/App.tsx
@@ -22,13 +22,13 @@ import { useBlockForMaintenance } from 'features/maintenance/useMaintenance'
 import { AppNavigationContainer } from 'features/navigation/NavigationContainer'
 import { SearchWrapper } from 'features/search/pages/SearchWrapper'
 import { ABTestingProvider } from 'libs/ABTesting'
+import { idCheckAnalytics } from 'libs/analytics'
 import { useCampaignTracker } from 'libs/campaign'
 import CodePushProvider from 'libs/codepush/CodePushProvider'
 import { env } from 'libs/environment'
 import { errorMonitoring } from 'libs/errorMonitoring'
 import { GeolocationWrapper } from 'libs/geolocation'
 import { activate } from 'libs/i18n'
-import { idCheckAnalytics } from 'libs/idCheckAnalytics'
 import { idCheckRetentionClient } from 'libs/idCheckRetentionClient'
 import { useStartBatchNotification } from 'libs/notifications'
 import { queryClient } from 'libs/queryClient'
