+++ src/App.web.tsx
@@ -14,10 +14,10 @@ import { AuthWrapper } from 'features/auth/AuthContext'
 import { AsyncErrorBoundaryWithoutNavigation } from 'features/errors/pages/AsyncErrorBoundary'
 import { FavoritesWrapper } from 'features/favorites/pages/FavoritesWrapper'
 import { AppNavigationContainer } from 'features/navigation/NavigationContainer'
+import { idCheckAnalytics } from 'libs/analytics'
 import { env } from 'libs/environment'
 import { errorMonitoring } from 'libs/errorMonitoring'
 import { activate } from 'libs/i18n'
-import { idCheckAnalytics } from 'libs/idCheckAnalytics'
 import { idCheckRetentionClient } from 'libs/idCheckRetentionClient'
 import { queryClient } from 'libs/queryClient'
 import { SafeAreaProvider } from 'libs/react-native-save-area-provider'
