+++ src/features/bookings/components/OnGoingBookingsList.tsx
@@ -6,8 +6,7 @@ import styled from 'styled-components/native'
 
 import { EndedBookingsSection } from 'features/bookings/pages/EndedBookingsSection'
 import { useFunctionOnce } from 'features/offer/services/useFunctionOnce'
-import { analytics } from 'libs/analytics'
-import { isCloseToBottom } from 'libs/analytics.utils'
+import { analytics, isCloseToBottom } from 'libs/analytics'
 import { Separator } from 'ui/components/Separator'
 import { TAB_BAR_COMP_HEIGHT } from 'ui/theme'
 import { ColorsEnum, getSpacing, Typo } from 'ui/theme'
