+++ src/features/bookings/pages/BookingDetails.tsx
@@ -18,8 +18,7 @@ import { contentHeight } from 'features/bookings/components/ThreeShapesTicket.co
 import { BookingProperties, getBookingProperties } from 'features/bookings/helpers'
 import { UseNavigationType, UseRouteType } from 'features/navigation/RootNavigator'
 import { useFunctionOnce } from 'features/offer/services/useFunctionOnce'
-import { analytics } from 'libs/analytics'
-import { isCloseToBottom } from 'libs/analytics.utils'
+import { analytics, isCloseToBottom } from 'libs/analytics'
 import { SeeItineraryButton } from 'libs/itinerary/components/SeeItineraryButton'
 import useOpenItinerary from 'libs/itinerary/useOpenItinerary'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
