+++ src/features/home/components/OffersModule.tsx
@@ -14,8 +14,7 @@ import { OfferTile, ModuleTitle, SeeMore } from 'features/home/atoms'
 import { SearchParametersFields, DisplayParametersFields } from 'features/home/contentful'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
 import { useFunctionOnce } from 'features/offer/services/useFunctionOnce'
-import { analytics } from 'libs/analytics'
-import { isCloseToEndHorizontal } from 'libs/analytics.utils'
+import { analytics, isCloseToEndHorizontal } from 'libs/analytics'
 import { formatDates, formatDistance, parseCategory, getDisplayPrice } from 'libs/parsers'
 import { SearchHit, parseSearchParameters } from 'libs/search'
 import { ColorsEnum, LENGTH_M, RATIO_HOME_IMAGE, Spacer } from 'ui/theme'
