+++ src/features/home/components/RecommendationModule.tsx
@@ -11,8 +11,7 @@ import styled from 'styled-components/native'
 import { OfferTile, ModuleTitle } from 'features/home/atoms'
 import { DisplayParametersFields } from 'features/home/contentful'
 import { useFunctionOnce } from 'features/offer/services/useFunctionOnce'
-import { analytics } from 'libs/analytics'
-import { isCloseToEndHorizontal } from 'libs/analytics.utils'
+import { analytics, isCloseToEndHorizontal } from 'libs/analytics'
 import { formatDates, formatDistance, parseCategory, getDisplayPrice } from 'libs/parsers'
 import { SearchHit } from 'libs/search'
 import { ColorsEnum, Spacer } from 'ui/theme'
