--- src/libs/analytics.native.test.ts
@@ -1,56 +1,0 @@
-export {}
-
-const mockLogEvent = jest.fn()
-
-jest.mock('@react-native-firebase/analytics', () => () => ({
-  logEvent: mockLogEvent,
-}))
-
-const { AnalyticsEvent, validateAnalyticsEvent, analytics } = jest.requireActual('./analytics')
-
-describe('analytics', () => {
-  describe('AnalyticsEvent', () => {
-    it('actual members should respect Firebase naming rules', () => {
-      for (const member of Object.values(AnalyticsEvent)) {
-        const isValidMember = validateAnalyticsEvent(member as string)
-        if (!isValidMember) {
-          fail(`${member} is not valid`)
-        }
-      }
-    })
-    it('fake member should fail the test when it is too long', () => {
-      expect(validateAnalyticsEvent('a'.repeat(39))).toBeTruthy()
-      expect(validateAnalyticsEvent('a'.repeat(40))).toBeTruthy()
-      expect(validateAnalyticsEvent('a'.repeat(41))).toBeFalsy()
-    })
-    it('fake member should fail the test when its first letter is not alphabetic', () => {
-      expect(validateAnalyticsEvent('!test')).toBeFalsy()
-      expect(validateAnalyticsEvent('1test')).toBeFalsy()
-      expect(validateAnalyticsEvent('_test')).toBeFalsy()
-    })
-    it('fake member should fail the test when it contains a non alpha numeric char', () => {
-      expect(validateAnalyticsEvent('Event-Testé')).toBeFalsy()
-    })
-    it('fake member should fail the test when it includes a "-"', () => {
-      expect(validateAnalyticsEvent('Event-Test')).toBeFalsy()
-    })
-    it('fake member should fail the test when it starts a reserved prefix', () => {
-      expect(validateAnalyticsEvent('firebase_test')).toBeFalsy()
-      expect(validateAnalyticsEvent('google_test')).toBeFalsy()
-      expect(validateAnalyticsEvent('ga_test')).toBeFalsy()
-      expect(validateAnalyticsEvent('test_firebase_test')).toBeTruthy()
-      expect(validateAnalyticsEvent('test_googletest')).toBeTruthy()
-      expect(validateAnalyticsEvent('test_ga_test')).toBeTruthy()
-    })
-  })
-
-  describe('logEvent', () => {
-    it('should cast offerId and bookingId from number to string', () => {
-      analytics.logBookingConfirmation(123456, 789)
-      expect(mockLogEvent).toHaveBeenCalledWith(AnalyticsEvent.BOOKING_CONFIRMATION, {
-        offerId: '123456',
-        bookingId: '789',
-      })
-    })
-  })
-})
