+++ src/libs/analytics/__mocks__/analytics.ts
@@ -1,17 +1,11 @@
-import { FirebaseAnalyticsTypes } from '@react-native-firebase/analytics'
-
-import { analytics as actualAnalytics } from '../analytics'
-
-type AnalyticsReturn = Partial<FirebaseAnalyticsTypes.Module>
-
-export const firebaseAnalytics: jest.Mocked<AnalyticsReturn> = {
-  logEvent: jest.fn(),
-  logLogin: jest.fn(),
-  logScreenView: jest.fn(),
-  setAnalyticsCollectionEnabled: jest.fn(),
-}
+import {
+  analytics as actualAnalytics,
+  idCheckAnalytics as actualIdCheckAnalytics,
+} from '../analytics'
 
 export const analytics: typeof actualAnalytics = {
+  disableCollection: jest.fn(),
+  enableCollection: jest.fn(),
   logAccessExternalOffer: jest.fn(),
   logAllModulesSeen: jest.fn(),
   logAllTilesSeen: jest.fn(),
@@ -40,24 +34,25 @@ export const analytics: typeof actualAnalytics = {
   logConsultWhyAnniversary: jest.fn(),
   logConsultWithdrawal: jest.fn(),
   logDiscoverOffers: jest.fn(),
-  logOpenLocationSettings: jest.fn(),
-  logOpenNotificationSettings: jest.fn(),
   logHasActivateGeolocFromTutorial: jest.fn(),
   logHasAddedOfferToFavorites: jest.fn(),
-  logHasChangedPassword: jest.fn(),
   logHasAppliedFavoritesSorting: jest.fn(),
-  logHasSkippedTutorial: jest.fn(),
+  logHasChangedPassword: jest.fn(),
   logHasRefusedCookie: jest.fn(),
+  logHasSkippedTutorial: jest.fn(),
   logHelpCenterContactSignupConfirmationEmailSent: jest.fn(),
+  logIdCheck: jest.fn(),
+  logLocationToggle: jest.fn(),
+  logLogin: jest.fn(),
   logLogout: jest.fn(),
   logMailTo: jest.fn(),
   logNoSearchResult: jest.fn(),
   logNotificationToggle: jest.fn(),
-  logOpenExternalUrl: jest.fn(),
   logOfferSeenDuration: jest.fn(),
-  logIdCheck: jest.fn(),
+  logOpenExternalUrl: jest.fn(),
+  logOpenLocationSettings: jest.fn(),
+  logOpenNotificationSettings: jest.fn(),
   logProfilScrolledToBottom: jest.fn(),
-  logLocationToggle: jest.fn(),
   logProfilSignUp: jest.fn(),
   logRecommendationModuleSeen: jest.fn(),
   logReinitializeFilters: jest.fn(),
@@ -72,4 +67,27 @@ export const analytics: typeof actualAnalytics = {
   logSignUpLessThanOrEqualTo13: jest.fn(),
   logUseFilter: jest.fn(),
   setUserId: jest.fn(),
+}
+
+export const idCheckAnalytics: typeof actualIdCheckAnalytics = {
+  cameraUnavailable: jest.fn(),
+  cancelSignUp: jest.fn(),
+  endCheckTokens: jest.fn(),
+  externalLink: jest.fn(),
+  fileSizeExceeded: jest.fn(),
+  getJouveToken: jest.fn(),
+  getLicenceToken: jest.fn(),
+  hasValidSession: jest.fn(),
+  idDocumentAcquisitionType: jest.fn(),
+  identityError: jest.fn(),
+  idValid: jest.fn(),
+  invalidAge: jest.fn(),
+  invalidDate: jest.fn(),
+  invalidDocument: jest.fn(),
+  invalidTwice: jest.fn(),
+  missingDocument: jest.fn(),
+  permissionsBlocked: jest.fn(),
+  processCompleted: jest.fn(),
+  startCheckTokens: jest.fn(),
+  wrongSideDocument: jest.fn(),
 }
