+++ src/libs/analytics/analytics.native.test.ts
@@ -1,0 +1,20 @@
+import { AnalyticsEvent } from './events'
+
+const mockLogEvent = jest.fn()
+jest.mock('@react-native-firebase/analytics', () => () => ({
+  logEvent: mockLogEvent,
+}))
+
+const { analytics } = jest.requireActual('./analytics')
+
+describe('analytics', () => {
+  describe('logEvent', () => {
+    it('should cast offerId and bookingId from number to string', () => {
+      analytics.logBookingConfirmation(123456, 789)
+      expect(mockLogEvent).toHaveBeenCalledWith(AnalyticsEvent.BOOKING_CONFIRMATION, {
+        offerId: '123456',
+        bookingId: '789',
+      })
+    })
+  })
+})
