+++ src/libs/analytics/utils.test.ts
@@ -1,6 +1,6 @@
 import { NativeScrollEvent } from 'react-native'
 
-import { isCloseToBottom } from './analytics.utils'
+import { isCloseToBottom } from './utils'
 
 describe('[Analytics utils]', () => {
   const nativeEventMiddle = {
